Update search-the-audit-log-in-security-and-compliance.md
Came accross an issue with a customer about the longer retention. I confimed with the Exchange Engineering Team that this feature isn't retroactive. You won't be able to pull out data up to 365 days prior to the date you were added to the Private Program Preview. I think it must be important to clarify that portion.